move resolver to exported util function
